Move the regression dataframe helper to conftest
